kuch fixes and refactoring kri h discuss krte badme
