+++ pkg/middlware/auth.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/Kaushik1766/chain-upi-gin/db"
 	"github.com/Kaushik1766/chain-upi-gin/pkg/utils"
 	"github.com/gin-gonic/gin"
 )
@@ -27,20 +26,6 @@ func Authenticate() gin.HandlerFunc {
 		}
 		ctx.Set("uid", parsedToken.UID)
 		ctx.Set("upi", parsedToken.UpiHandle)
-		ctx.Next()
-	}
-}
-
-func Verify() gin.HandlerFunc {
-	return func(ctx *gin.Context) {
-		upi, _ := ctx.Get("uid")
-		walletAddress := ctx.Query("walletAddress")
-		err := db.VerifyWallet(walletAddress, upi.(string))
-		if err != nil {
-			ctx.AbortWithStatusJSON(http.StatusBadRequest, gin.H{
-				"error": err.Error(),
-			})
-		}
 		ctx.Next()
 	}
 }
